changed interpreting of map tiles

+++ include/Map.h
@@ -16,7 +16,7 @@ public:
     void draw(int x , int y);
 
     Sprite * elements[37] ;
-    int width , height;
+    double width , height;
 
     void reloadGUI();
